Modificacion de tipo id en CAtalogo Ropa de Long a int. Correccion de clase repositorio, En un principio se hizo de tipo interface pero seguimos la guia del profe y la hicimos tipo clase

+++ src/main/java/com/duoc/RopaStore/models/CatalogoRopa.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor // Indicamos que la clase tendra unc ontructor con todos los parametros
 @NoArgsConstructor // Indicamos que la clase tendra un constructor sin parametros
 public class CatalogoRopa {
-    private Long idRopa;
+    private int idRopa;
     private String categoriaPrenda;
     private String marcaPrenda;
     private String tipoPrenda;
